# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TMJConnect API
  version: 1.0.0
  description: |
    HIPAA-compliant orofacial pain management platform API.
    Supports patients, providers, and admin roles.
  contact:
    name: TMJConnect Engineering
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Report:
      description: Patient-to-provider health report
      type: object
      required: [id, patient_id, urgency, description, status, flagged, submitted_at]
      properties:
        id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        provider_id: { type: [string, "null"], format: uuid, description: "NULL if provider deleted" }
        urgency: { type: string, enum: [routine, concerning, urgent] }
        pain_level: { type: [integer, "null"], minimum: 0, maximum: 10 }
        description: { type: string, description: "Patient's symptom narrative (sanitised)" }
        photo_url: { type: [string, "null"], format: uri }
        period_start: { type: [string, "null"], format: date-time }
        period_end: { type: [string, "null"], format: date-time }
        summary_data:
          $ref: '#/components/schemas/ReportSummaryData'
        patient_notes: { type: [string, "null"], description: "Additional patient notes (sanitised)" }
        status: { type: string, enum: [submitted, viewed, reviewed, responded] }
        flagged: { type: boolean }
        submitted_at: { type: string, format: date-time }
        viewed_at: { type: [string, "null"], format: date-time }
        reviewed_at: { type: [string, "null"], format: date-time }
        authored_by_user_id:
          type: [string, "null"]
          format: uuid
          description: "Who authored the row. Equals patient_id for patient-submitted reports; equals provider_id for on-behalf-of reports."
        authored_by_role:
          type: string
          enum: [patient, provider]
          description: "Provenance flag. 'provider' = filed on-behalf-of via /providers/patients/:id/reports."
...
    ReportRequest:
      description: Provider-to-patient nudge asking for a report. Fulfilled when a linked report is filed.
      type: object
      required: [id, patient_id, prompt, status, created_at]
      properties:
        id: { type: string, format: uuid }
        provider_id: { type: [string, "null"], format: uuid }
        patient_id: { type: string, format: uuid }
        prompt: { type: string, maxLength: 2000 }
        status: { type: string, enum: [pending, fulfilled, dismissed] }
        fulfilled_report_id: { type: [string, "null"], format: uuid }
        fulfilled_at: { type: [string, "null"], format: date-time }
        dismissed_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
    ReportSummaryData:
      description: Auto-generated stats snapshot at time of report submission
      type: object
      properties:
        avg_pain_7d: { type: [number, "null"], description: "7-day average pain level" }
        symptom_count_7d: { type: [integer, "null"], description: "Symptom logs in last 7 days" }
        exercises_completed_7d: { type: [integer, "null"], description: "Exercises completed in last 7 days" }
      additionalProperties: true
    ReportInboxItem:
      description: Summary row in the provider report inbox (description truncated to 200 chars)
      type: object
      required: [id, patient_id, urgency, status, flagged, submitted_at, patient_first_name, patient_last_name]
      properties:
        id: { type: string, format: uuid }
        patient_id: { type: string, format: uuid }
        urgency: { type: string, enum: [routine, concerning, urgent] }
        status: { type: string, enum: [submitted, viewed, reviewed, responded] }
        pain_level: { type: [integer, "null"] }
        description_preview: { type: string, maxLength: 200, description: "First 200 chars of description" }
        flagged: { type: boolean }
        submitted_at: { type: string, format: date-time }
        patient_first_name: { type: string }
        patient_last_name: { type: string }
    ReportResponse:
      description: Provider text response to a report
      type: object
      required: [id, message, responded_at]
      properties:
        id: { type: string, format: uuid }
        message: { type: string, description: "Response visible to patient (sanitised)" }
        internal_notes:
          type: [string, "null"]
          description: "Provider-only notes. NEVER returned in patient-facing responses."
        responded_at: { type: string, format: date-time }
...
    PageMeta:
      description: Offset-based pagination metadata
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, description: "Current page number" }
        limit: { type: integer, description: "Items per page" }
        total: { type: integer, description: "Total matching items" }
        totalPages: { type: integer, description: "Total pages" }
...
paths:
...
  /reports:
    post:
      tags: [Reports]
      summary: Submit a health report to provider (patient)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: Idempotency-Key
          in: header
          schema: { type: string, maxLength: 64 }
          description: Optional. Prevents duplicate submissions on retry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id, urgency, description]
              properties:
                provider_id: { type: string, format: uuid }
                urgency: { type: string, enum: [routine, concerning, urgent] }
                pain_level: { type: [integer, "null"], minimum: 0, maximum: 10 }
                description: { type: string, maxLength: 10000 }
                photo_url: { type: [string, "null"], format: uri }
                period_start: { type: [string, "null"], format: date-time }
                period_end: { type: [string, "null"], format: date-time }
                patient_notes: { type: [string, "null"], maxLength: 5000 }
      responses:
        '201': { description: Report submitted. Provider notified. }
        '200': { description: Idempotent replay (report already exists). }
        '403': { description: Not linked to provider. }
  /reports/inbox:
    get:
      tags: [Reports]
      summary: Provider report inbox (sorted by urgency)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [submitted, viewed, reviewed, responded] }
        - name: patient_id
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
        - name: urgency
          in: query
          schema: { type: string, enum: [routine, concerning, urgent] }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReportInboxItem' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
  /reports/{id}:
    get:
      tags: [Reports]
      summary: Get report detail (both roles)
      security: [{ bearerAuth: [] }]
      description: |
        Patient view: no internal_notes in responses.
        Provider view: includes internal_notes, auto-marks report as viewed.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      report: { $ref: '#/components/schemas/Report' }
                      responses:
                        type: array
                        items: { $ref: '#/components/schemas/ReportResponse' }
  /reports/{id}/respond:
    post:
      tags: [Reports]
      summary: Provider responds to a report
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 5000 }
                internal_notes:
                  type: [string, "null"]
                  maxLength: 5000
                  description: Provider-only notes. Never visible to patients.
      responses:
        '201': { description: Response created. Patient notified. }
  /reports/{id}/review:
    patch:
      tags: [Reports]
      summary: Mark report as reviewed (provider)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Report marked as reviewed. Patient notified. }
  /reports/{id}/flag:
    patch:
      tags: [Reports]
      summary: Toggle report flag (provider)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Flag toggled }
  /reports/requests:
    get:
      tags: [Reports]
      summary: List report requests (role-aware)
      description: Patients see their own pending requests. Providers see their sent requests, with optional filters.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, fulfilled, dismissed] }
        - name: patient_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReportRequest' }
  /reports/requests/{id}:
    delete:
      tags: [Reports]
      summary: Dismiss a pending report request (either party)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204': { description: Dismissed }
        '403': { description: Not allowed to dismiss this request }
        '409': { description: Request is not pending }

  # ─── Provider-only: clinical notes + report requests + on-behalf-of reports ───
...
  /providers/patients/{patientId}/reports:
    post:
      tags: [Providers]
      summary: File a report on behalf of a patient
      description: >
        Creates a report with `authored_by_role='provider'`. Use when the patient
        can't file directly (phone call, in-person visit, escalation).
        Pass `fulfilling_request_id` to close out a pending report_request atomically.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: patientId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [urgency, description]
              properties:
                urgency: { type: string, enum: [routine, concerning, urgent] }
                pain_level:
                  type: integer
                  minimum: 0
                  maximum: 10
                  nullable: true
                description: { type: string, maxLength: 10000 }
                photo_url: { type: string, format: uri, nullable: true }
                period_start: { type: string, format: date-time, nullable: true }
                period_end: { type: string, format: date-time, nullable: true }
                patient_notes: { type: string, maxLength: 5000, nullable: true }
                fulfilling_request_id: { type: string, format: uuid }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Report' }

  # ──────────────────────────── LINKING ────────────────────────────
...
  /reports/mine:
    get:
      tags: [Reports]
      summary: Patient — list own submitted reports
      description: Reverse-chronological list of the calling patient's own reports, with offset pagination.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: urgency
          in: query
          schema: { type: string, enum: [routine, concerning, urgent] }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Patient's own reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Report' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
        '403': { description: Not a patient role }

  # ──────────────────────────── PROVIDERS (additions) ───────────────────────────
...
  /admin/reports:
    get:
      tags: [Admin]
      summary: List all reports across all providers (admin oversight)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: urgency, in: query, schema: { type: string, enum: [routine, concerning, urgent] } }
        - { name: status, in: query, schema: { type: string, enum: [submitted, viewed, reviewed, responded] } }
        - { name: unanswered_over_hours, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Report' }
                  meta: { $ref: '#/components/schemas/PageMeta' }
...